Initial commit
Doesn't do anywhere near what's needed to be a Norton Guide reader yet. What
it does do is test a couple of key ideas and also goes as far as reading the
header of a guide. I'm tracking this early on because I want to get it into
a net-based cloud repo so I can hack on it from various machines as and when
the fancy takes me (and time allows).