Try and make insert-file-contents-literally a little less hungry
Either I'm using it wrong, or there's a difference in how
insert-file-contents-literally works and what the documentation says. If I
use it to constantly replace the content of a buffer with data from a file
it seems to fill up the buffer rather than replace the content.

I need to isolate this and better understand it.