Better test for the RLE marker
Seems that checking if a character is equal to the RLE marker (which is 255)
doesn't quite cut it in all situations in Emacs. Better that I do the check
to see if a character is equal to 255 or it's equal to whatever

  (decode-char 'eight-bit 255)

is. In testing, the former works fine in my code if I'm messing with a
hard-coded string for testing purposes, eg:

  (eg-expand-string "Hello\377\10World")

but when reading from the buffer that loads the guide data from the guide,
the test against the decode is required. I'm still not 100% sure why -- I'll
dig and read up on this at some point.