Try and ensure that multibyte doesn't get involved
I'm still struggling to get this to read more than a couple of entries into
the test guide before it falls over. I think most of my problems come from
the handling of multibyte characters. This change makes it fail in a
less-terrible way so I seem to be getting closer.