Don't test for eof in eg-skip-entry
At this point we're not on the start of a guide anyway so the test isn't
quite valid. "Fixing" this does introduce an older problem (that of
correctly detecting eof) but at least I now know what it is I need to fix
and how not to fix it.

Trying to mimic file I/O in buffers is a little more tricky (for me) than
I'd anticipated. Trying to write this code like it's all the other NG code
I've written might not be the best idea.