Getting dynamic see-also menus going
This isn't quite working yet. While I've got the menu showing fine, with the
right content, when there are see-also entries to be had, I can't quite get
the menu to not show when that isn't the case. I can't even seem to have it
be inactive when that's the case. I think I need to learn a little more
about easymenu.